# Remove read-ahead for small files